+++ front-page.php
@@ -25,7 +25,7 @@ add_action( 'wp_enqueue_scripts', __NAMESPACE__ . '\enqueue_front_page_assets' )
  * @return void
  */
 function enqueue_front_page_assets() {
-	wp_enqueue_style( 'ktc_front_page_css', CHILD_URL . '/assets/dist/css/front-page.min.css', array(), '1.6.1' );
+	wp_enqueue_style( 'ktc_front_page_css', CHILD_URL . '/assets/dist/css/front-page.min.css', array(), '1.6.2' );
 }
 
 genesis();