+++ template-viewing-history.php
@@ -1,0 +1,17 @@
+<?php
+/**
+ * Viewing History Template
+ *
+ * Template Name: Viewing History
+ *
+ * @package     KnowTheCode\ViewingHistory
+ * @since       1.5.1
+ * @author      hellofromTonya
+ * @link        https://UpTechLabs.io
+ * @license     GNU General Public License 2.0+
+ */
+namespace KnowTheCode\ViewingHistory;
+
+remove_action( 'genesis_before_loop', 'genesis_do_breadcrumbs' );
+remove_all_actions( 'genesis_entry_header' );
+genesis();